+++ scripts/events/combat.js
@@ -15,7 +15,7 @@ function collectMacros(entity) {
 function collectTokenMacros(token, pass, distance, target) {
     let triggers = [];
     if (token.actor) {
-        let effects = actorUtils.getEffects(token.actor);
+        let effects = actorUtils.getEffects(token.actor, {includeItemEffects: true});
         effects.forEach(effect => {
             let macroList = collectMacros(effect);
             if (!macroList.length) return;